feat: Replace client binary with dedicated example
